+++ test/acceptance/setup-terraform/main.tf
@@ -80,32 +80,6 @@ resource "aws_cloudwatch_log_group" "log_group" {
   tags = var.tags
 }
 
-// Policy that allows execution of remote commands in ECS tasks.
-resource "aws_iam_policy" "execute_command" {
-  name   = "ecs-execute-command-${local.suffix}"
-  path   = "/"
-  policy = <<EOF
-{
-  "Version": "2012-10-17",
-  "Statement": [
-    {
-      "Effect": "Allow",
-      "Action": [
-        "ssmmessages:CreateControlChannel",
-        "ssmmessages:CreateDataChannel",
-        "ssmmessages:OpenControlChannel",
-        "ssmmessages:OpenDataChannel"
-      ],
-      "Resource": [
-        "*"
-      ]
-    }
-  ]
-}
-EOF
-
-}
-
 resource "hcp_hvn" "server" {
   hvn_id         = "hvn-${local.suffix}"
   cloud_provider = "aws"